Move stylesheets to assets folder
